% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAI-helpers.R
\name{calculate_ti_bounds}
\alias{calculate_ti_bounds}
\title{TI censoring bounds calculation
Internal helper for censoring transformation that you shouldn't use unless
you know what you are doing. Based on the table in the supplement of
Yang's censoring paper: https://doi.org/10.1101/2022.08.25.22279230}
\usage{
calculate_ti_bounds(y, x)
}
\arguments{
\item{y}{Vector of log-scale post-vacciation titer values}

\item{x}{Vector of log-scale pre-vaccination titer values.}
}
\value{
A tibble containining three numeric columns: `z`, the log-scale
titer increase, `z_l`, the lower bound of `z`, and `z_u`, the upper
bound of `z`.
}
\description{
TI censoring bounds calculation
Internal helper for censoring transformation that you shouldn't use unless
you know what you are doing. Based on the table in the supplement of
Yang's censoring paper: https://doi.org/10.1101/2022.08.25.22279230
}
