% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-de-chunker.R
\name{detect_os}
\alias{detect_os}
\title{Detect operating system}
\usage{
detect_os(verbose = TRUE)
}
\arguments{
\item{verbose}{should a message be printed when the function runs?}
}
\value{
`.Platform$OS.type`
}
\description{
We need to run a shell command for the model (de)chunker, and the way we have
to do that is different or Windows vs unix-based OS. So this function will
detect the OS that an individual is using. If it isn't Windows or Unix, the
user will get an error and the (de)chunker won't work.
}
\examples{
detect_os()
}
