% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAI-helpers.R
\name{format_hai_data}
\alias{format_hai_data}
\title{Convert HAI data to brms censoring format
Main function that takes a post titer data column as a string
(and optionally a pretiter column if you want the outcome to be titer
increase), calculates the censoring bounds, and then converts to the
correct brms format.}
\usage{
format_hai_data(
  data,
  post_titer,
  pre_titer = NULL,
  log_scale = TRUE,
  increase = FALSE,
  log_out = log_scale
)
}
\arguments{
\item{data}{A data frame}

\item{post_titer}{A string, the name of the column in `data` containing the
post-vaccination titer values.}

\item{pre_titer}{A string, the name of the column in `data` containing
the pre-vaccination titer values. Optional, only required if titer increase
should be calculated.}

\item{log_scale}{Boolean, TRUE if the titers are already on the log scale
and FALSE if the titer columns are on the natural scale. Defaults to TRUE.
Currently the post_titer and pre_titer columns must both be on the same
scale.}

\item{increase}{Boolean, TRUE if the outcome should be titer increase, and
FALSE if the outcome should be post-vaccination titer. Defaults to FALSE.}

\item{log_out}{Boolean, TRUE if the outcomes should be returned on the
log scale, FALSE if the outcomes should be returned on the natural scale.
Defaults to the scale of the input variables.}
}
\value{
A data frame with three additional columns named y, y2, and c. Note
that if you already have these columns in the data frame, they will be
overwritten in the result and you might get unexpected behavior.
}
\description{
Convert HAI data to brms censoring format
Main function that takes a post titer data column as a string
(and optionally a pretiter column if you want the outcome to be titer
increase), calculates the censoring bounds, and then converts to the
correct brms format.
}
