% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-de-chunker.R
\name{get_directory_path}
\alias{get_directory_path}
\title{Extract directory from file path}
\usage{
get_directory_path(file_path)
}
\arguments{
\item{file_path}{a character vector of file paths}
}
\value{
a character vector of directorys in which the files would be saved.
}
\description{
Utility function that, given a file path, returns the entire file path
without the final component that has a file extension. The point is to
get the full directory path that a file should live in.
}
