% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAI-helpers.R
\name{hai_to_log_scale}
\alias{hai_to_log_scale}
\title{hai_to_log_scale()
Convert an HAI titer to log scale using the log2(x / 5) transform.}
\usage{
hai_to_log_scale(x)
}
\arguments{
\item{x}{the natural scale titer}
}
\value{
log2(x / 5)
}
\description{
hai_to_log_scale()
Convert an HAI titer to log scale using the log2(x / 5) transform.
}
