% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAI-helpers.R
\name{hai_to_natural_scale}
\alias{hai_to_natural_scale}
\title{hai_to
Convert a logged HAI titer back to natural scale using 5 * 2 ^ x transform.}
\usage{
hai_to_natural_scale(x)
}
\arguments{
\item{x}{the log scale titer vector}
}
\value{
5 * 2 ^ x
}
\description{
hai_to
Convert a logged HAI titer back to natural scale using 5 * 2 ^ x transform.
}
