% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAI-helpers.R
\name{hai_trans}
\alias{hai_trans}
\title{transform HAI titers
Function to convert between log and natural scale for HAI titers.}
\usage{
hai_trans(x, dir = c("log", "natural"))
}
\arguments{
\item{x}{HAI titers on either the log or natural scale.}

\item{dir}{either 'log' if the titers should be transformed to the log scale
or 'natural' if the titers should be transformed to the natural scale.
Partial matching is supported.}
}
\value{
The transformed titer values, a numeric vector
}
\description{
transform HAI titers
Function to convert between log and natural scale for HAI titers.
}
