% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-de-chunker.R
\name{model_cat}
\alias{model_cat}
\title{Model dechunker}
\usage{
model_cat(chunk_directories, destination_file)
}
\arguments{
\item{chunk_directories}{character vector of length n.}

\item{destination_file}{character vector also of length n.}
}
\value{
invisibly, the last shell call result.
}
\description{
The two arguments are both character vectors that should be equal length.
Each directory specified in the vector "chunk_directories" should be a
directory created by invoking model_cat() that contains the individual chunks
from a splitted file. This function recombines then and saves the combined
file to the file name in the same position in the vector "destination_file".
This invokes the bash command "cat" which is available natively on both
windows and MacOS. If you're on a different OS you might need to install it
or something, I don't know. You can learn how the "cat" command works here:
\url{https://web.archive.org/web/20250228174620/https://www.man7.org/linux/man-pages/man1/cat.1.html}
or by googling it.
}
