% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-de-chunker.R
\name{model_split}
\alias{model_split}
\title{Model Chunker}
\usage{
model_split(files_to_chunk, destination_directories)
}
\arguments{
\item{files_to_chunk}{character vector of length n.}

\item{destination_directories}{character vector also of length n.}
}
\value{
invisibly, the last result of a shell command to execute.
}
\description{
The two arguments are both character vectors that should be equal length. It
takes each file specified in the vector "files_to_chunk", separates it into
smaller chunks with a fixed memory size, and saves all of the chunks to the
DIRECTORY PATH at the same position specified in "destination_directories".
This invokes the bash command "split" which is available natively on both
windows and MacOS. If you're on a different OS you might need to install it
or something, I don't know. You can learn how the "split" command works here:
\url{https://web.archive.org/web/20250228173733/https://man7.org/linux/man-pages/man1/split.1.html}
or by googling it. Note that the backslash character indicates an escape
character in an R string (\url{https://en.wikipedia.org/wiki/Escape_character}).
}
