% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_strain_names.R
\name{replace_strain_names}
\alias{replace_strain_names}
\title{Replace strain names with a different format}
\usage{
replace_strain_names(x, from = "analysis", to = "short", drop = TRUE)
}
\arguments{
\item{x}{A vector of strain names.}

\item{from}{Format of the names in the vector `x`. Should be one of
"analysis", "full", or "short". See ["handelgroup_strain_names"] for the
allowed names in each of the formats. If you want to transform a strain
that is not currently in the strain list, you will need to add it and
submit a PR to `hgp`!}

\item{to}{Format of the returned names. Should be one of "short", "full",
"analysis", or "subtype".}

\item{drop}{If TRUE, levels of the returned factor variable are dropped. If
FALSE, the level set of the factor will still contain every strain in
["handelgroup_strain_names"], which is typically not desirable.}
}
\value{
A factor vector of the same length as `x`.
}
\description{
Using the standardized list of handelgroup strain names from
["handelgroup_strain_names"], pass in a vector of strain names of arbitrary
length and return the same sequence of names in a different format.
}
\examples{
replace_strain_names("CA/09", from = "short", to = "analysis")

dat <- data.frame(s = c("CA/09", "MI/15"), x = c(1, 2))
transform(
    dat,
    s_long = replace_strain_names(s, from = "short", to = "analysis")
)
}
